% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fastq_type.R
\name{check_fastq_type}
\alias{check_fastq_type}
\title{check_fastq_type}
\usage{
check_fastq_type(x, chemistry = NULL, n = 50)
}
\arguments{
\item{x}{full path to fastq file}

\item{chemistry}{in development. specify cellRanger chemistry type, v1 - v5, if known for more precise detection.}

\item{n}{number of rows of fastq file to use}
}
\value{
character string. One of barcode, sample, or read.
}
\description{
Determine scRNA-seq fastq type. Fastq can be one of: UMI+Barcode, Sample Index, or Reads. Only works with short-read sequencing.
}
