% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gate_plot.R
\name{create_gate_plot}
\alias{create_gate_plot}
\title{Create Gate Visualization}
\usage{
create_gate_plot(
  sce,
  marker1,
  marker2,
  x_boundary,
  y_boundary,
  population_name,
  assay_name
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{marker1}{First marker name (x-axis)}

\item{marker2}{Second marker name (y-axis)}

\item{x_boundary}{Numeric value for vertical gate position}

\item{y_boundary}{Numeric value for horizontal gate position}

\item{population_name}{Character string for plot title}

\item{assay_name}{Character string specifying which assay to use}
}
\value{
ggplot object
}
\description{
Create Gate Visualization
}
