% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_gene_format.R
\name{detect_gene_format}
\alias{detect_gene_format}
\title{detect_gene_format}
\usage{
detect_gene_format(gene_ids)
}
\arguments{
\item{x}{character vector of gene IDs}
}
\value{
format of gene IDs
}
\description{
given a vector of gene ids, automatically detect the gene fomrat from standard formats (ensembl, entrez, refseq, hgnc, uscs). Only provides with format declaration if all genes IDs match that format.
}
