% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_doublets.R
\name{plot_doublets}
\alias{plot_doublets}
\title{plot_doublets}
\usage{
plot_doublets(
  sce,
  plot_by,
  doublet_cols = c("cxds_score", "bcds_score", "hybrid_score")
)
}
\arguments{
\item{sce}{sce object}

\item{plot_by}{string. single variable to plot by}

\item{doublet_cols}{character vector. name of doublet score columns. default should work if using scdcs.}
}
\value{
plot
}
\description{
plot doublet scores. designed to work with scds doublet scores calculated in run_sc_qc pipeline.
}
\examples{
plot_doublets(sce = sce, plot_by = params$sample_col)
}
