% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manual_pairwise.R
\name{plot_manual_pairwise}
\alias{plot_manual_pairwise}
\title{plot_manual_pairwise}
\usage{
plot_manual_pairwise(
  sce,
  features,
  thresholds,
  altExp_name = "PROTEIN",
  exprs_value = "logcounts"
)
}
\arguments{
\item{sce}{sce object}

\item{features}{characteor vector of the 2 features to plot}

\item{thresholds}{character vectors of the thresholds for each of the 2 features in the same order as features}

\item{altExp_name}{sce object alt exp name to use}

\item{exprs_values}{sce object assay to use}
}
\value{
list of sce objects
}
\description{
pairwise plots with manual gates. include reports of percentage of cells in each gated population.
}
\examples{
threshold$nk$lognorm <- c(CD3 = 0.25, CD56 = 0)
gating$nk$lognorm <- plot_manual_pairwise(sce,
                                          features = c("CD3", "CD56"),
                                          thresholds = threshold$nk$lognorm,
                                          exprs_value = "logcounts")
}
