% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mrh.R
\name{plot_mrh}
\alias{plot_mrh}
\title{plot_sample_qc}
\usage{
plot_mrh(sce, plot_by, mit_cutoff)
}
\arguments{
\item{sce}{sce object}

\item{plot_by}{character vector; variables to plot by}

\item{mit_cutoff}{numeric value; a horizontal line will be added at this threshold.}
}
\value{
list of plots
}
\description{
plot standard quality control metrics of mitochondrial, ribosomal, and hemoglobin gene counts. designed to work with run_sc_qc pipeline only.
}
\examples{
plot_mrh(sce, plot_by = c("sample", "annot"), mit_cutoff = 5)
}
