% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_qc.R
\name{plot_sample_qc}
\alias{plot_sample_qc}
\title{plot_sample_qc}
\usage{
plot_sample_qc(sce, plot_by)
}
\arguments{
\item{sce}{sce object}

\item{plot_by}{character vector; variables to plot by}
}
\value{
list of plots
}
\description{
plot standard quality control metrics of total UMI and total detected genes per sample
}
\examples{
plot_sample_qc(sce, plot_by = c("sample", "annot"))
}
