% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_clr.R
\name{sce_clr}
\alias{sce_clr}
\title{sce_clr
Apply centered log-ratio transformation to SingleCellExperiment object}
\usage{
sce_clr(x, altExp_name = "PROTEIN")
}
\arguments{
\item{x}{A SingleCellExperiment object containing the data to transform}

\item{altExp_name}{Character string specifying the name of the alternative
experiment containing the count data (default: "PROTEIN")}
}
\value{
A modified SingleCellExperiment object with:
  \itemize{
    \item Original counts stored in "rawcounts" assay
    \item CLR-transformed data stored in both "counts" and "logcounts" assays
  }
}
\description{
Performs centered log-ratio (CLR) transformation on count data stored in an
alternative experiment of a SingleCellExperiment object. The function adds
pseudocounts before transformation and performs quality checks on the results.
}
\details{
The function performs the following steps:
1. Adds pseudocounts using the pseudo_count function
2. Applies CLR transformation using the compositions package
3. Performs quality checks on row and column names
4. Stores the results in appropriate assay slots
}
\examples{
# Assuming 'sce' is a SingleCellExperiment object with protein data
sce <- sce_clr(sce, altExp_name = "PROTEIN")

}
