% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_for_seurat.R
\name{sce_for_seurat}
\alias{sce_for_seurat}
\title{sce_for_seurat}
\usage{
sce_for_seurat(x, keep_dim = NULL)
}
\arguments{
\item{x}{sce object}

\item{keep_dim}{characteor vector of dimensions to keep}
}
\value{
seurat object
}
\description{
converts RNA and PROTEIN assays from single cell experiment to seurat format. reduced dimensions are removed by default unless specified to be kept.
}
\examples{
sce_for_seurat(keep_dim = "tcell_totalVI")
}
