% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_for_seurat2.R
\name{sce_for_seurat2}
\alias{sce_for_seurat2}
\title{sce_for_seurat2}
\usage{
sce_for_seurat2(
  x,
  rawcounts = "rawcounts",
  keep_additional_assays = NULL,
  keep_dim = NULL,
  altExp = NULL
)
}
\arguments{
\item{x}{sce object}

\item{rawcounts}{name of assay containing rawcounts}

\item{keep_additional_assays}{character vector of any additional assays - other than rawcounts - to include in seurat object}

\item{keep_dim}{character vector of dimensions to keep}

\item{altExp}{which alt experiments to include in seurat object}
}
\value{
seurat object. count layer is the rawcounts layer specified.
}
\description{
converts assays from single cell experiment to seurat format. Designed for h5 export for scvi. Reduced dimensions are removed by default unless specified to be kept.
}
\examples{
sce_for_seurat(keep_dim = "tcell_totalVI")
}
