% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_h5.R
\name{seurat_to_h5}
\alias{seurat_to_h5}
\title{seurat_to_h5}
\usage{
seurat_to_h5(x, output_dir_prefix, assays = c("RNA", "PROTEIN"))
}
\arguments{
\item{x}{seurt object}

\item{output_dir_prefix}{# full path to output directory + any prefix. Will append assays and h5.Seurat to the file}

\item{altExp_name}{sce object alt exp name to  to export. one file will be exported per each experiment.}
}
\value{
exports h5seurat and h5ad files from a seurat object
}
\description{
Converts RNA and PROTEIN experiments from a seurat object into an h5ad file for import to Python.
}
\examples{
seurat_to_h5(seurat, "path/to/output/directory/output_prefix")
}
