% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sce.R
\name{split_sce}
\alias{split_sce}
\title{split_sce}
\usage{
split_sce(sce, by)
}
\arguments{
\item{sce}{sce object}

\item{by}{character string; variable to split by}
}
\value{
list of sce objects
}
\description{
split sce object into multiple objects by variable
}
\examples{
split_sce(sce = sce, by = "sample_id")
}
