% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractScSampleId.R
\name{extractScSampleId}
\alias{extractScSampleId}
\title{extractScSampleId}
\usage{
extractScSampleId(x, auto = F, type = NULL, regex = NULL)
}
\arguments{
\item{x}{filename to extract sample Id from}

\item{auto}{binary. if true, will attempt to extract common scrnaseq sample Id formats from filename}

\item{type}{character vector. specify which common sample IDs from common single cell formats to extract. Options include: sraSample, sraRun, sraExperiment, sraProject, sraAccession}

\item{regex}{user specified regex to extract sampleId. If provided alongside type, any match to common regex type(s) or user-input regex will be used}
}
\value{
sample ID
}
\description{
extract sample Id from common single cell formats
}
