% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol.R
\name{get_symbol}
\alias{get_symbol}
\title{get_symbol}
\usage{
get_symbol(gene_ids, from_type = "auto", add_attribute = NULL, missing = T)
}
\arguments{
\item{gene_ids}{character vector of gene IDs (required)}

\item{from_type}{Optional. specify format of gene IDs. One of: ensembl, entrez, refseq, hgnc, ucsc, symbol. If not provided, will automatically try and detect format.}

\item{add_attribute}{Optional. additional gene attributes to add. use listAttributes for ensembl & hsapiens_gene_ensembl to see what attributes are available.}

\item{missing}{Default = TRUE. Binary. If TRUE, any gene IDs not found in the ensembl database will still be included in the results. Symbol will be the original gene ID instead.}
}
\value{
dataframe of gene symbol and any additional attributes
}
\description{
query a vector of gene ids using biomaRt ensembl database and return symbol. Optionally return additional attributes.
}
