% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol.R
\name{get_symbol}
\alias{get_symbol}
\title{get_symbol}
\usage{
get_symbol(gene_ids, from_type = "auto", add_attribute = NULL)
}
\arguments{
\item{gene_ids}{character vector of gene IDs (required)}

\item{from_type}{(optional) specify format of gene IDs. One of:  If not provided, will automatically try and detect format.}

\item{add_attribute}{(optional) additional gene attributes to add. use useMart("ensembl", dataset = "hsapiens_gene_ensembl") %>% listAttributes to see what attributes are available.}
}
\value{
dataframe of gene symbol and any additional attributes
}
\description{
query a vector of gene ids using biomaRt ensembl database and return symbol. Optionally return additional attributes.
}
