% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mrh.R
\name{plot_mrh}
\alias{plot_mrh}
\title{plot_mrh}
\usage{
plot_mrh(sce, plot_by, mit_cutoff, include = "mrh")
}
\arguments{
\item{sce}{sce object}

\item{plot_by}{character vector; variables to plot by}

\item{mit_cutoff}{numeric value; a horizontal line will be added at this threshold.}

\item{include}{character vector containing m, r, and/or h. If m is in vector, mitochondrial genes will be plotted. If r, ribosomal genes. If h, hemoglobin genes.}
}
\value{
list of plots
}
\description{
plot standard quality control metrics of mitochondrial, ribosomal, and hemoglobin gene counts. designed to work with run_sc_qc pipeline only.
}
\examples{
plot_mrh(sce, plot_by = c("sample", "annot"), mit_cutoff = 5, include = "mrh")
}
