% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_count.R
\name{pseudo_count}
\alias{pseudo_count}
\title{pseudo_count
Add pseudocounts to numeric data}
\usage{
pseudo_count(x)
}
\arguments{
\item{x}{Numeric vector containing the data}
}
\value{
Numeric vector with pseudocounts added
}
\description{
Adds half of the smallest non-zero value in the dataset to each value. This is
useful for handling zero values before log-ratio transformations.
}
\examples{
data <- c(0, 1, 2, 3, 0, 4)
pseudo_count(data)

}
